doctr_mod/requirements.txt
python-doctr[torch]
opencv-python==4.11.0.86
pillow==11.2.1
pytesseract==0.3.13
pyyaml==6.0.2
numpy==2.0.2
pandas==2.3.0
pdf2image==1.17.0
pymupdf
PyPDF2==3.0.1
tqdm==4.67.1
office365-rest-python-client
openpyxl==3.1.2
easyocr
